RF: separated test_odf.py to test_odf.py and test_peaks.py for consistency with the new peaks.py file
